Joining and sending code for the network to execute should now be operational

+++ example/Fibonacci.java
@@ -1,16 +1,14 @@
 public class Fibonacci {
-
-public static void main (String[] args) {
-    int f = 0;
-    int g = 1;
+  public static void main(String[] args) {
+    long f = 0;
+    long g = 1;
 
     for (int i = 1; i <= 101; i++) {
-        System.out.print(f + " ");
-        f = f + g;
-        g = f - g;
-    } 
+      System.out.print(f + " ");
+      f = f + g;
+      g = f - g;
+    }
 
     System.out.println();
-}
-
+  }
 }
